refactor(app): self-contain import dialog with its own create mutation and query select
Move project.create into ImportProjectDialog and derive entries via useQuery select.
