+++ apps/app/src/components/projects/projects-sidebar-group.tsx
@@ -1,4 +1,4 @@
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
 import { useRouteContext } from "@tanstack/react-router";
 import {
   SidebarGroup,
@@ -11,31 +11,18 @@ import {
 } from "@vibest/ui/components/sidebar";
 import { FolderCode, FolderPlus } from "lucide-react";
 import { useState } from "react";
-import { toast } from "sonner";
 
 import { ImportProjectDialog } from "@/components/projects/import-project-dialog";
 
 /** The Projects sidebar section: lists imported projects and drives the import flow. */
 export function ProjectsSidebarGroup() {
   const { orpcQueryUtils } = useRouteContext({ from: "__root__" });
-  const queryClient = useQueryClient();
   const [importOpen, setImportOpen] = useState(false);
 
-  // The only writer is the create mutation below, which invalidates on success.
+  // The only writer is the import dialog's create mutation, which invalidates on success.
   const projects = useQuery({
     ...orpcQueryUtils.project.list.queryOptions(),
     staleTime: Infinity,
-  });
-
-  const importProject = useMutation({
-    mutationFn: (path: string) => orpcQueryUtils.project.create.call({ path }),
-    onSuccess: () => {
-      setImportOpen(false);
-      return queryClient.invalidateQueries({ queryKey: orpcQueryUtils.project.list.key() });
-    },
-    onError: (error) => {
-      toast.error(`Failed to import project: ${error.message}`);
-    },
   });
 
   return (
@@ -58,13 +45,7 @@ export function ProjectsSidebarGroup() {
         </SidebarMenu>
       </SidebarGroupContent>
 
-      {importOpen && (
-        <ImportProjectDialog
-          onClose={() => setImportOpen(false)}
-          onImport={importProject.mutate}
-          importing={importProject.isPending}
-        />
-      )}
+      {importOpen && <ImportProjectDialog onClose={() => setImportOpen(false)} />}
     </SidebarGroup>
   );
 }
